Move class UserPowerLimit.
